Update sequential.cpp
Changed double to float

+++ homework/sequential.cpp
@@ -4,9 +4,9 @@
 #include "matrix.h"
 #include "sequential.h"
 
-pc::matrix<int> floydWarhsallSeq(pc::matrix<double> &dist){
+pc::matrix<int> floydWarhsallSeq(pc::matrix<float> &dist){
 
-    double tr;
+    float tr;
     int pr;
     int d = dist.getSize();
     pc::matrix<int> pred = pc::matrix<int>(d);
@@ -52,8 +52,8 @@ pc::matrix<int> floydWarhsallSeq(pc::matrix<double> &dist){
 // dx,dy: start index of submatrix d
 // ax,ay: start index of submatrix a
 // bx,by: start index of submatrix b
-void floyd(pc::matrix<double> &dist, pc::matrix<int> &pred, int dx, int dy, int ax, int ay, int bx, int by, int s) {
-    double tr;
+void floyd(pc::matrix<float> &dist, pc::matrix<int> &pred, int dx, int dy, int ax, int ay, int bx, int by, int s) {
+    float tr;
     int pr;
     for(int h=0; h<s;h++){
         for(int i=0; i<s;i++){
@@ -75,13 +75,13 @@ void floyd(pc::matrix<double> &dist, pc::matrix<int> &pred, int dx, int dy, int 
     }
 }
 
-pc::matrix<int> floydWarhsallSquared(pc::matrix<double> &dist, int s){
+pc::matrix<int> floydWarhsallSquared(pc::matrix<float> &dist, int s){
     int n = dist.getSize();
 
     pc::matrix<int> pred = pc::matrix<int>(n);
 
 
-    if(((int)((double)n/(double)s))%2 == 0) {
+    if(((int)((float)n/(float)s))%2 == 0) {
         std::cout << "ERROR! work only with a odd division of matrix";
         return pc::matrix<int>(0);
     }
